requirements.txt
discord.py==1.7.3
python-dotenv==0.17.1
aiohttp==3.7.4.post0
discord-py-slash-command==2.3.2
APScheduler==3.7.0
pandas==1.3.1
tabulate==0.8.9
motor==2.5.0
pymongo[srv]==3.12.0
coloredlogs==15.0.1